refactor: 메서드 네이밍 수정
