+++ src/main/java/roomescape/controller/dto/ReservationCreateRequest.java
@@ -1,11 +1,13 @@
-package roomescape.domain;
+package roomescape.controller.dto;
+
+import roomescape.domain.Reservation;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
 
 public record ReservationCreateRequest(String name, LocalDate date, LocalTime time) {
 
-    public Reservation toEntity(int id) {
+    public Reservation toReservation(int id) {
         return new Reservation(id, name, date, time);
     }
 }
