+++ src/main/java/roomescape/domain/Reservation.java
@@ -5,10 +5,10 @@ import java.time.LocalTime;
 
 public class Reservation {
 
-    private int id;
-    private String name;
-    private LocalDate date;
-    private LocalTime time;
+    private final int id;
+    private final String name;
+    private final LocalDate date;
+    private final LocalTime time;
 
     public Reservation(int id, String name, LocalDate date, LocalTime time) {
         this.id = id;
@@ -19,14 +19,6 @@ public class Reservation {
 
     public int getId() {
         return id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public LocalDate getDate() {
-        return date;
     }
 
     public LocalTime getTime() {
